Update docker infrastructure and dependencies

+++ castor.php
@@ -2,14 +2,24 @@
 
 use Castor\Attribute\AsTask;
 
+use function Castor\guard_min_version;
 use function Castor\import;
 use function Castor\io;
 use function Castor\notify;
+use function Castor\variable;
+use function docker\about;
+use function docker\build;
+use function docker\docker_compose_run;
+use function docker\generate_certificates;
+use function docker\pull;
+use function docker\up;
+
+guard_min_version('0.18.0');
 
 import(__DIR__ . '/.castor');
 
 /**
- * @return array<string, mixed>
+ * @return array{project_name: string, root_domain: string, extra_domains: string[], php_version: string}
  */
 function create_default_variables(): array
 {
@@ -19,33 +29,25 @@ function create_default_variables(): array
     return [
         'project_name' => $projectName,
         'root_domain' => "{$projectName}.{$tld}",
-        'php_version' => $_SERVER['DS_PHP_VERSION'] ?? '8.2',
-        'project_directory' => '.',
-    ];
-}
-
-/**
- * @return array<string, mixed>
- */
-function create_default_environment(): array
-{
-    return [
-        'BUILDER_VERSION' => 'latest',
-        'FRONTEND_VERSION' => 'latest',
-        'ROUTER_VERSION' => 'latest',
+        'extra_domains' => [
+            "www.{$projectName}.{$tld}",
+        ],
+        'php_version' => '8.3',
     ];
 }
 
 #[AsTask(description: 'Builds and starts the infrastructure, then install the application (composer, yarn, ...)')]
 function start(bool $build = false): void
 {
-    infra\generate_certificates(false);
+    io()->title('Starting the stack');
+
+    generate_certificates(force: false);
     if ($build) {
-        infra\build();
+        build();
     } else {
-        infra\pull();
+        pull();
     }
-    infra\up();
+    up(profiles: ['default']); // We can't start worker now, they are not installed
     cache_clear();
     install();
     migrate();
@@ -57,22 +59,62 @@ function start(bool $build = false): void
     about();
 }
 
-#[AsTask(description: 'Installs the application (composer, yarn, ...)', namespace: 'app')]
+#[AsTask(description: 'Installs the application (composer, yarn, ...)', namespace: 'app', aliases: ['install'])]
 function install(): void
 {
-    docker_compose_run('composer install -n --prefer-dist --optimize-autoloader');
+    io()->title('Installing the application');
+
+    $basePath = variable('root_dir');
+
+    if (is_file("{$basePath}/composer.json")) {
+        io()->section('Installing PHP dependencies');
+        docker_compose_run('composer install -n --prefer-dist --optimize-autoloader');
+    }
+    if (is_file("{$basePath}/yarn.lock")) {
+        io()->section('Installing Node.js dependencies');
+        docker_compose_run('yarn install --frozen-lockfile');
+    } elseif (is_file("{$basePath}/package.json")) {
+        io()->section('Installing Node.js dependencies');
+
+        if (is_file("{$basePath}/package-lock.json")) {
+            docker_compose_run('npm ci');
+        } else {
+            docker_compose_run('npm install');
+        }
+    }
+    if (is_file("{$basePath}/importmap.php")) {
+        io()->section('Installing importmap');
+        docker_compose_run('bin/console importmap:install');
+    }
+
     qa\install();
 }
 
-#[AsTask(description: 'Clears the application cache', namespace: 'app')]
+#[AsTask(description: 'Clear the application cache', namespace: 'app', aliases: ['cache-clear'])]
 function cache_clear(): void
 {
-    docker_compose_run('rm -rf var/cache/ && bin/console cache:warmup');
+    io()->title('Clearing the application cache');
+
+    docker_compose_run('rm -rf var/cache/');
+    // On the very first run, the vendor does not exist yet
+    if (is_dir(variable('root_dir') . '/vendor')) {
+        docker_compose_run('bin/console cache:warmup');
+    }
 }
 
-#[AsTask(description: 'Migrates database schema', namespace: 'app:db')]
+#[AsTask(description: 'Migrates database schema', namespace: 'app:db', aliases: ['migrate'])]
 function migrate(string $env = 'dev'): void
 {
+    io()->title('Migrating the database schema');
+
     docker_compose_run('bin/console doctrine:database:create --if-not-exists --env=' . $env);
-    docker_compose_run('bin/console doctrine:migration:migrate -n --allow-no-migration --env=' . $env);
+    docker_compose_run('bin/console doctrine:migration:migrate -n --allow-no-migration --all-or-nothing --env=' . $env);
+}
+
+#[AsTask(description: 'Loads fixtures', namespace: 'app:db', aliases: ['fixture'])]
+function fixtures(): void
+{
+    io()->title('Loads fixtures');
+
+    docker_compose_run('bin/console doctrine:fixture:load -n');
 }
